[feat] 반복되는 VC, VM 통합
