+++ HilingualPresentation/Sources/Presentation/Feed/FeedCell/FeedListViewController.swift
@@ -1,0 +1,44 @@
+//
+//  FeedListViewController.swift
+//  HilingualPresentation
+//
+//  Created by 조영서 on 8/25/25.
+//
+
+import UIKit
+import Foundation
+import Combine
+
+public final class FeedListViewController: BaseUIViewController<FeedViewModel> {
+
+    // MARK: - Properties
+    private let feedCellView = FeedCellView()
+    private let input = FeedViewModel.Input()
+    
+    // MARK: - Lifecycle
+    public override func loadView() {
+        self.view = feedCellView
+    }
+
+    public override func viewDidLoad() {
+        super.viewDidLoad()
+        bindViewModel()
+        input.reload.send(())
+    }
+
+    // MARK: - Bindings
+    private func bindViewModel() {
+        guard let output = viewModel?.transform(input: input) else { return }
+
+        Publishers.CombineLatest(output.feeds, output.haveFollowing)
+            .receive(on: RunLoop.main)
+            .sink { [weak self] (feeds, haveFollowing) in
+                guard let self else { return }
+                self.feedCellView.apply(
+                    items: feeds,
+                    followingState: haveFollowing
+                )
+            }
+            .store(in: &cancellables)
+    }
+}
