--- HilingualPresentation/Sources/Presentation/Feed/FollowingFeed/FollowingFeedViewController.swift
@@ -1,37 +1,0 @@
-//
-//  FollowingFeedViewController.swift
-//  HilingualPresentation
-//
-//  Created by 조영서 on 8/19/25.
-//
-
-import UIKit
-import Foundation
-
-public final class FollowingFeedViewController: BaseUIViewController<FollowingFeedViewModel> {
-
-    private let feedCellView = FeedCellView()
-    private let input = FollowingFeedViewModel.Input()
-
-    public override func loadView() {
-        self.view = feedCellView
-    }
-
-    public override func viewDidLoad() {
-        super.viewDidLoad()
-        bindViewModel()
-        input.reload.send(())
-    }
-
-    private func bindViewModel() {
-        let output = viewModel?.transform(input: input)
-        
-        output?.feeds
-            .receive(on: RunLoop.main)
-            .sink { [weak self] (feeds, haveFollowing) in
-                guard let self = self else { return }
-                self.feedCellView.apply(items: feeds, haveFollowing: haveFollowing)
-            }
-            .store(in: &cancellables)
-    }
-}
